Ask The Expert
This repository is for sample scripts created for educational purposes for the Cisco "Ask The Expert" webinar series.

These scripts are not meant to demonstrate Python scripting best practices or the most elegant solutions to specific problems. They are simply meant for those starting out with Python and with DNA Center Platform API's, and the hope is that this will spark some ideas for how regular network engineers can begin using Intent API's on a daily basis to simplify their most common tasks.